cloud_function/inference_function/requirements.txt
functions-framework
google-cloud-aiplatform
google-api-python-client
protobuf